Moved the unsplash API in the Gallery component
- Data normalization
- Removed previous component
