+++ src/components/Gallery.jsx
@@ -1,27 +1,55 @@
 // Hooks
-import { useState } from "react";
+import { useState, useEffect } from "react";
 // Components
+import Spinner from "./Spinner";
 import GalleryTab from "./GalleryTab";
-import GalleryTabUnsplash from "./GalleryUnsplashTab";
 import Button from "./Button";
 // Gallery data
 import { gallery } from "../gallery";
-// Unsplash Icon
+
+// Unsplash API key
+const unsplash_access = "8sTE-SoLOIgv4NGA7NoyQ-PLy8N24rpsWfZZK71AqPA";
+// Unsplash icon
 import { unsplash } from "../assets";
+// Unsplash API plugin
+import { createApi } from "unsplash-js";
+// Unsplash API
+const api = createApi({
+	// Don't forget to set your access token here!
+	// See https://unsplash.com/developers
+	accessKey: unsplash_access,
+});
 
 // TODO: Ajouter un hash à l'URL et ouvrir l'image concernée au load
 
 const Gallery = () => {
 	const [isUnsplash, setIsUnsplash] = useState(false);
-	const [isUnsplashLoaded, setIsUnsplashLoaded] = useState(false);
+	// Unsplash API data state
+	const [data, setPhotosResponse] = useState(null);
 
 	const disableUnsplash = () => {
 		setIsUnsplash(false);
 	};
 	const enableUnsplash = () => {
 		setIsUnsplash(true);
-		setIsUnsplashLoaded(true);
 	};
+
+	useEffect(() => {
+		// Unsplash API Call
+		api.users
+			.getPhotos({
+				username: "maxime_dore",
+				page: 1,
+				perPage: 30,
+				orderBy: "popular", // latest, oldest, popular, views, downloads
+			})
+			.then((result) => {
+				setPhotosResponse(result);
+			})
+			.catch((e) => {
+				console.log("Une erreur s'est produite en appelant l'API Unsplash.", e);
+			});
+	}, []);
 
 	return (
 		<section id="galerie" className="bg-black">
@@ -58,7 +86,9 @@ const Gallery = () => {
 					</div>
 					{/* Unsplash gallery */}
 					<div className={`gallery__tab ${isUnsplash ? "" : "tab--hidden"}`}>
-						<div className="wrapper">{isUnsplashLoaded && <GalleryTab source="unsplash" />}</div>
+						<div className="wrapper">
+							{data === null ? <Spinner /> : <GalleryTab source={data.response.results} type="unsplash" />}
+						</div>
 					</div>
 					{/* TODO: Intégrer l'API Instagram pour un tab de feed custom */}
 				</div>
